% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canmap-package.R
\docType{package}
\name{canmap-package}
\alias{canmap}
\alias{canmap-package}
\title{canmap: Standard Canadian geography shapefiles}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jesse Tweedle \email{jesse.tweedle@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
