% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_pos.R
\docType{data}
\name{code_pos}
\alias{code_pos}
\title{Directory of metadata character positions for Canadian geographic datasets}
\format{A data frame.
\describe{
  \item{code_type}{The type of code, (e.g., \code{geo_level})}
  \item{start_chr}{Start position of the code in the filename (e.g., 2 = start of \code{geo_level})}
  \item{stop_chr}{End position of the code in the filename (e.g., 4 = end of \code{geo_level})}
}}
\source{
\url{https://www150.statcan.gc.ca/n1/pub/92-196-x/92-196-x2016001-eng.htm}
}
\usage{
code_pos
}
\description{
Code positions of geographic metadata in filenames
}
\keyword{datasets}
