% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_geography.R
\name{download_geography}
\alias{download_geography}
\title{Download and unzip Canadian shapefiles}
\usage{
download_geography(geo_path)
}
\arguments{
\item{geo_path}{X}
}
\value{
A tibble of shapefiles along with metadata, including
\code{filepath}, \code{ext}, \code{type}, \code{path},
\code{ref_date}, \code{geo_code}, \code{geo_level},
\code{file_type}, \code{format}, \code{projection}, \code{geo_coverage},
and \code{language}.
}
\description{
Download and unzip Canadian shapefiles from Statistics Canada
}
\examples{
\donttest{
download_geography("http://www12.statcan.gc.ca/census-recensement/2011/
geo/bound-limit/files-fichiers/2016/lpr_000a16a_e.zip")
}
}
