% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geo_info}
\alias{geo_info}
\title{Get metadata associated with geographic filename}
\usage{
geo_info(filename)
}
\arguments{
\item{filename}{Name of geography file}
}
\value{
A one-row tibble of metadata associated with the filename
of Statistics Canada's shapefiles.
}
\description{
Get metadata associated with geographic filename
}
\examples{
\donttest{
geo_info("lpr_000a16a_e")
}
}
