% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geoinfo.R
\name{get_geoinfo}
\alias{get_geoinfo}
\title{Get metadata associated with geographic filename.}
\usage{
get_geoinfo(filename)
}
\arguments{
\item{filename}{Name of geography file}
}
\value{
A one-row tibble of metadata associated with the filename
of Statistics Canada's shapefiles.
}
\description{
Get metadata associated with geographic filename.
}
\examples{
\donttest{
get_geoinfo("lpr_000a16a_e")
}
}
