% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_geography}
\alias{download_geography}
\title{Download and unzip Canadian shapefiles}
\usage{
download_geography(geo_path, geo_dir = NULL, overwrite = FALSE)
}
\arguments{
\item{geo_path}{Full path/url to zipped shapefile}

\item{geo_dir}{Directory to download to --- may not keep this.}

\item{overwrite}{Download and unzip even if geography
already exists in \code{geo_dir}}
}
\value{
A tibble of shapefiles along with metadata, including
\code{filepath}, \code{size}, \code{path},
\code{ref_date}, \code{geo_code}, \code{geo_level},
\code{file_type}, \code{format}, \code{projection},
and \code{language}.
}
\description{
Download and unzip Canadian shapefiles from Statistics Canada
}
\examples{
\donttest{
download_geography("http://www12.statcan.gc.ca/census-recensement/2011/
geo/bound-limit/files-fichiers/2016/lpr_000a16a_e.zip")
}
}
